Extract match to own file
