+++ lib/match.test.js
@@ -1,7 +1,6 @@
 "use strict";
 
 var assert = require("@sinonjs/referee").assert;
-require("jsdom-global")();
 var samsam = require("./samsam");
 
 describe("match", function () {
